Capture form and ribbon design
i made a app that connects solidworks ( a common 3d modeling software used by engineers to make parts) with excel. users can quickly capture part measurements from solidworks, list them in an excel sheet, modify the values, and then update the solidworks model with these new values. this process makes it easy to adjust parts and reduces the time needed for custom modifications.

first i thought it would be easy use python to create an app extract the values from solidworks. but it was not easy at all i had to do alot of research to make this functional i went down a youtube rabbit hole of part automation in solidworks found one video of a simple program that worked with excel and i reverse engineered it along with the help of grimoire a chatgpt built specifically for code. it was alot of trial and error and i think i had to delete the codebase many times to get it right i think yesterday was the peak i had 16 errors all across all my program files. i also learned not to copy paste blindly code to actually think about what im putting down does it make logical sense etc. now i have a functional app and it can only get better from here. the leap was to connect to solidworks and i was able to do that and modify dimensions of a part right from a excel spreadsheet!